fix(footer): stabilization of e2e tests
