+++ packages/web-components/tests/e2e-storybook/cypress/support/index.js
@@ -7,6 +7,8 @@
 
 import '@percy/cypress';
 import serializeDOM from '@percy/dom';
+import translationData from '../fixtures/translation.json';
+import countryData from '../fixtures/countrylist.json';
 
 import './commands';
 
@@ -78,5 +80,9 @@ beforeEach(() => {
         isDataLayerReady: true,
       },
     };
+
+    // Set sessionStorage data
+    win.sessionStorage.setItem('dds-translation-us-en', JSON.stringify(translationData));
+    win.sessionStorage.setItem('dds-countrylist-en-us', JSON.stringify(countryData));
   });
 });
